requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
flask-restx==1.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
gunicorn==20.1.0
marshmallow==3.19.0
marshmallow-dataclass==8.5.11
PyJWT==2.6.0
pytest==7.2.1
python-dotenv==0.21.1
requests==2.28.2
SQLAlchemy==1.4.45
WTForms==3.0.1